requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
python-multipart
passlib[bcrypt]
email-validator
python-jose[cryptography]
fastapi-users[sqlalchemy]
python-dotenv
requests
wikipedia
html5lib
beautifulsoup4
sentence-transformers>=2.2.0 # Atau versi terbaru
scikit-learn # Untuk cosine_similarity jika tidak pakai util dari sentence-transformers